# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Silvercoin Loan Management API
  description: |
    Backend API for the Silvercoin loan management app. Authenticated routes use `Authorization: Bearer <accessToken>`.
    Login is email + password only (no SMS OTP).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local (default PORT from `.env`)
  - url: http://localhost:5000
    description: Fallback if PORT unset
...
paths:
...
  /api/employee/profile:
    get:
      tags: [Employee]
      summary: Current employee profile plus tenantLogoUrl and tenantName from parent admin when available
      operationId: employeeProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
        '404':
          $ref: '#/components/responses/NotFoundMessage'
    patch:
      tags: [Employee]
      summary: Update display name only
      operationId: employeePatchProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/employee/loans:
    post:
      tags: [Employee]
      summary: Create draft loan
      operationId: employeeCreateLoan
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  loanId:
                    type: string
    get:
      tags: [Employee]
      summary: List my loans
      operationId: employeeListLoans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
  /api/employee/loans/{loanId}:
    get:
      tags: [Employee]
      summary: Get my loan
      operationId: employeeGetLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          $ref: '#/components/responses/NotFoundMessage'
    patch:
      tags: [Employee]
      summary: Update draft fields (draft or submitted only)
      operationId: employeePatchLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerName:
                  type: string
                ownerMobile:
                  type: string
                ownerEmail:
                  type: string
                  format: email
                  description: Optional; empty clears stored value
                area:
                  type: string
                marketValue:
                  type: string
                geoLocation:
                  $ref: '#/components/schemas/GeoLocation'
                loanAmount:
                  oneOf:
                    - type: number
                    - type: string
                emiPerMonth:
                  oneOf:
                    - type: number
                    - type: string
                totalMonths:
                  oneOf:
                    - type: integer
                    - type: string
                transferCount:
                  type: integer
                  minimum: 0
                  maximum: 999
                  description: Times property was transferred to a descendant (employee-entered)
                otherLoanOnProperty:
                  type: boolean
                  description: Whether another active loan exists on the same property
      responses:
        '200':
          description: OK; registryTag present once eligibility is met
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  registryTag:
                    type: string
                    nullable: true
        '400':
          description: Loan not editable in current status
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/upload-photo:
    post:
      tags: [Employee]
      summary: Upload property photos (repeat field `photos`, or legacy single `photo`; appended; max 15 total per loan)
      operationId: employeeUploadPhoto
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more image files (same field name repeated)
                photo:
                  type: string
                  format: binary
                  description: Legacy single-file upload (same behavior as one `photos` part)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  propertyImageKeys:
                    type: array
                    items:
                      type: string
                    description: All S3 keys for this loan after append
                  registryTag:
                    type: string
                    nullable: true
        '400':
          description: No files or over photo limit
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/registry-document:
    post:
      tags: [Employee]
      summary: Upload property registry paper (PDF or image); field name `document`
      operationId: employeeUploadRegistryDocument
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [document]
              properties:
                document:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  registryDocumentKey:
                    type: string
                  registryDocumentContentType:
                    type: string
                  registryTag:
                    type: string
                    nullable: true
        '400':
          description: Invalid type or not editable
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/upload-video:
    post:
      tags: [Employee]
      summary: Upload video (multipart field `video`). Mobile app records owner verification video with a 60s maximum duration (client-enforced).
      operationId: employeeUploadVideo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [video]
              properties:
                video:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/submit:
    post:
      tags: [Employee]
      summary: Submit draft (generates PDF to S3, status → submitted)
      operationId: employeeSubmitLoan
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  pdfKey:
                    type: string
        '400':
          description: Already submitted or missing required fields
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/agreement:
    post:
      tags: [Employee]
      summary: Upload agreement images/signatures; status → agreement_submitted
      operationId: employeeSubmitAgreement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                agreementDoc:
                  type: string
                  format: binary
                employeeSignature:
                  type: string
                  format: binary
                ownerSignature:
                  type: string
                  format: binary
                agreementId:
                  type: string
                agreementDate:
                  type: string
                witnessName:
                  type: string
                witnessMobile:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBody'
        '400':
          description: Invalid state
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/media-preview:
    get:
      tags: [Employee]
      summary: Presigned URLs for property images, video, and registry document
      operationId: employeeLoanMediaPreview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  propertyImageUrls:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                        url:
                          type: string
                          format: uri
                  videoUrl:
                    type: string
                    format: uri
                    nullable: true
                  registryDocumentUrl:
                    type: string
                    format: uri
                    nullable: true
                  registryDocumentContentType:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFoundMessage'
  /api/employee/loans/{loanId}/pdf:
    get:
      tags: [Employee]
      summary: Presigned URL for generated loan PDF
      operationId: employeeLoanPdfUrl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/loanId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: Time-limited S3 GET URL
        '404':
          description: Not found or PDF not generated yet
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/auth/login`
  parameters:
    loanId:
      name: loanId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: express-validator errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
...
    NotFoundMessage:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageBody'
  schemas:
    MessageBody:
      type: object
      properties:
        message:
          type: string
    ValidationErrorBody:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    User:
      type: object
      description: Stored user item without passwordHash in API responses
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [superadmin, admin, employee]
        isFirstLogin:
          type: boolean
        isActive:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        organizationLogoKey:
          type: string
          description: S3 key for admin organization logo (admins only)
...
    EmployeeProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            tenantLogoUrl:
              type: string
              format: uri
              nullable: true
            tenantName:
              type: string
              nullable: true
    GeoLocation:
      type: object
      description: Property GPS point; capturedAt is device time when the employee tapped capture (ISO 8601)
      properties:
        lat:
          type: number
        lng:
          type: number
        capturedAt:
          type: string
          format: date-time
          description: When coordinates were captured (new captures only; legacy loans may omit)
      required: [lat, lng]
    Loan:
      type: object
      description: MongoDB loan document (fields vary by status)
      additionalProperties: true
      properties:
        loanId:
          type: string
        employeeId:
          type: string
        adminId:
          type: string
        status:
          type: string
          description: draft | submitted | agreement_submitted | approved | active | rejected | completed
        createdAt:
          type: string
        updatedAt:
          type: string
        ownerName:
          type: string
        ownerMobile:
          type: string
        ownerEmail:
          type: string
          nullable: true
          description: Optional; used for notifications / QR email when present
        area:
          type: string
        marketValue:
          type: string
        geoLocation:
          $ref: '#/components/schemas/GeoLocation'
        loanAmount:
          type: string
        emiPerMonth:
          type: string
        totalMonths:
          type: integer
        transferCount:
          type: integer
          minimum: 0
          maximum: 999
          description: Times property was transferred to a descendant
        otherLoanOnProperty:
          type: boolean
          description: Whether another active loan exists on the same property
        ownerPhotoKey:
          type: string
          description: First property image key (legacy / mirror of propertyImageKeys[0])
        propertyImageKeys:
          type: array
          items:
            type: string
          description: S3 keys for property images (ordered)
        videoKey:
          type: string
        pdfKey:
          type: string
        agreementDocKey:
          type: string
        employeeSignatureKey:
          type: string
        ownerSignatureKey:
          type: string
        agreementId:
          type: string
        agreementDate:
          type: string
        witnessName:
          type: string
        witnessMobile:
          type: string
        approvedBy:
          type: string
        approvedAt:
          type: string
        loanStartDate:
          type: string
          description: Loan start date (YYYY-MM-DD). EMI schedule — first due 30 days after start, then each EMI 30 days after the previous (30, 60, 90… days from start)
        remainingMonths:
          type: integer
        rejectionReason:
          type: string
        upiQrKey:
          type: string
        registryTag:
          type: string
          description: Yearly display tag (e.g. 2026001) set once when owner, property, and loan fields are complete
          nullable: true
        registryDocumentKey:
          type: string
          description: S3 key for uploaded property registry PDF or scan
          nullable: true
        registryDocumentContentType:
          type: string
          nullable: true
...